+++ homepage/src/data/link-data.ts
@@ -1,0 +1,65 @@
+export const linksData = [
+  {
+    id: 'top',
+    name: 'Top',
+    main: true,
+  },
+  {
+    id: 'getting-started',
+    name: 'Getting started',
+    main: true,
+  },
+  {
+    id: 'install',
+    name: 'Install',
+  },
+  {
+    id: 'create-store',
+    name: 'Create a store',
+  },
+  {
+    id: 'use-store',
+    name: 'Use store',
+  },
+  {
+    id: 'recipes',
+    name: 'Recipes',
+    main: true,
+  },
+  {
+    id: 'create-and-use-store',
+    name: 'Create and use store',
+  },
+  {
+    id: 'subscribe',
+    name: 'Subscribe',
+  },
+  {
+    id: 'multi-value-store',
+    name: 'Multi value store',
+  },
+  {
+    id: 'reset-value',
+    name: 'Reset value',
+  },
+  {
+    id: 'session-storage',
+    name: 'sessionStorage',
+  },
+  {
+    id: 'new-local-storage',
+    name: 'New storage',
+  },
+  {
+    id: 'change-default-storagename',
+    name: 'Default storage name',
+  },
+  {
+    id: 'typescript',
+    name: 'Typescript',
+  },
+  {
+    id: 'react-hooks',
+    name: 'React hooks',
+  },
+];